+++ src/main/java/com/svetylkovo/rojo/annotations/Mapper.java
@@ -1,0 +1,13 @@
+package com.svetylkovo.rojo.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import java.util.function.Function;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Mapper {
+    Class<? extends Function<String, ?>> value();
+}