+++ src/test/java/com/svetylkovo/rojo/matcher/RojoBeanProcessorTest.java
@@ -1,12 +1,13 @@
 package com.svetylkovo.rojo.matcher;
 
-import com.svetylkovo.rojo.matcher.beans.*;
 import com.svetylkovo.rojo.exceptions.MissingDateFormatAnnotationException;
 import com.svetylkovo.rojo.exceptions.MissingRegexAnnotationException;
+import com.svetylkovo.rojo.matcher.beans.*;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
+import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 
 public class RojoBeanProcessorTest {
@@ -45,10 +46,10 @@ public class RojoBeanProcessorTest {
 
         MatchIterator it = new MatchIterator(processor.getMatcher(flagsTestString));
 
-        BeanIterator<NoFlagsBean> beanProcessor = new BeanIterator<>(it, processor);
-        beanProcessor.hasNext();
+        BeanIterator<NoFlagsBean> beanIterator = new BeanIterator<>(it, processor);
+        beanIterator.hasNext();
 
-        NoFlagsBean bean = beanProcessor.next();
+        NoFlagsBean bean = beanIterator.next();
         assertEquals("abc,def,", bean.getValue());
     }
 
@@ -59,11 +60,26 @@ public class RojoBeanProcessorTest {
 
         MatchIterator it = new MatchIterator(processor.getMatcher(flagsTestString));
 
-        BeanIterator<FlagsBean> beanProcessor = new BeanIterator<>(it, processor);
-        beanProcessor.hasNext();
+        BeanIterator<FlagsBean> beanIterator = new BeanIterator<>(it, processor);
+        beanIterator.hasNext();
 
-        FlagsBean bean = beanProcessor.next();
+        FlagsBean bean = beanIterator.next();
         assertEquals("abc,def,\nghi,jkl", bean.getValue());
+    }
+
+    @Test
+    public void mapperTest() {
+        RojoBeanProcessor<MapperBean> processor = new RojoBeanProcessor<>(MapperBean.class);
+        processor.processAnnotations();
+
+        MatchIterator it = new MatchIterator(processor.getMatcher("123"));
+
+        BeanIterator<MapperBean> beanIterator = new BeanIterator<>(it, processor);
+        beanIterator.hasNext();
+
+        MapperBean bean = beanIterator.next();
+
+        assertArrayEquals(new int[]{1,2,3}, bean.getNumbers());
     }
 
 }