+++ CHANGES.md
@@ -4,4 +4,5 @@
 - added forEach() method for direct group extraction using a lambda function up to 10 parameters
 - added asIterable() and asMatcherIterable()
 - added matchIterable() in RojoBeanMatcher
+- added @Mapper annotation for group mapping to the custom field type
 - removed asMatcherList()